Add GNU AGPL v3 License
Added the GNU Affero General Public License version 3 to the project. This will continue kaia's sole wish that no one profit from their code and their work. It does not fully disallow commercial use, but anything made derivative from this point forward on this repo will be required to have the same license as this project and their code must be open-sourced too, even if they try to charge for it.